move PrunedStore definitions into pruned.rs submodule
