+++ crates/wasmi/src/store/mod.rs
@@ -1,7 +1,10 @@
+mod pruned;
+
+pub use self::pruned::PrunedStore;
+use self::pruned::RestorePrunedWrapper;
 use crate::{
     collections::arena::{Arena, ArenaIndex, GuardedEntity},
     core::{
-        hint::unlikely,
         ElementSegment as CoreElementSegment,
         Fuel,
         Global as CoreGlobal,
@@ -38,7 +41,6 @@ use alloc::{boxed::Box, sync::Arc};
 use core::{
     any::{type_name, TypeId},
     fmt::{self, Debug},
-    mem,
     sync::atomic::{AtomicU32, Ordering},
 };
 
@@ -100,26 +102,6 @@ struct CallHookWrapper<T>(Box<dyn FnMut(&mut T, CallHook) -> Result<(), Error> +
 impl<T> Debug for CallHookWrapper<T> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "CallHook<{}>", type_name::<T>())
-    }
-}
-
-/// A wrapper used to restore a [`PrunedStore`].
-///
-/// This wrapper exists to provide a `Debug` impl so that `#[derive(Debug)]`
-/// works for [`Store`].
-#[allow(clippy::type_complexity)]
-#[derive(Clone)]
-struct RestorePrunedWrapper(Arc<dyn Send + Sync + Fn(&mut PrunedStore) -> &mut dyn TypedStore>);
-impl RestorePrunedWrapper {
-    /// Restores the [`PrunedStore`] and returns a reference to it via [`TypedStore`].
-    #[inline]
-    fn restore<'a>(&self, pruned: &'a mut PrunedStore) -> &'a mut dyn TypedStore {
-        (self.0)(pruned)
-    }
-}
-impl Debug for RestorePrunedWrapper {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "RestorePrunedWrapper")
     }
 }
 
@@ -194,148 +176,6 @@ pub struct Store<T> {
     id: TypeId,
     /// Used to restore a [`PrunedStore`] to a [`Store<T>`].
     restore_pruned: RestorePrunedWrapper,
-}
-
-/// A [`Store`] with a pruned `T`.
-#[derive(Debug)]
-#[repr(transparent)]
-pub struct PrunedStore {
-    /// The underlying [`Store`] with pruned type signature.
-    pruned: Store<Pruned>,
-}
-
-/// Placeholder type of `T` for a pruned `Store<T>`.
-#[derive(Debug)]
-pub struct Pruned;
-
-impl<'a, T> From<&'a mut Store<T>> for &'a mut PrunedStore {
-    #[inline]
-    fn from(store: &'a mut Store<T>) -> Self {
-        // Safety: the generic `Store<T>` has its `T` pruned here.
-        //
-        // - This is safe because we are operating on a `&mut Store<T>` thus it is just
-        //   a reference and since `Store<T>` and `Store<Pruned>` have the same size and API.
-        // - We make sure in `PrunedStore` to never access the typed parts of the original
-        //   `Store<T>` and check in the restoration process the type-ID of the target `T`.
-        // - `Store<T>` has the same size and alignment for all `T`.
-        unsafe { mem::transmute::<&'a mut Store<T>, &'a mut PrunedStore>(store) }
-    }
-}
-
-impl<T> Store<T> {
-    /// Prune the [`Store`] from `T` returning a [`PrunedStore`].
-    #[inline]
-    pub(crate) fn prune(&mut self) -> &mut PrunedStore {
-        self.into()
-    }
-}
-
-impl PrunedStore {
-    /// Returns a shared reference to the underlying [`StoreInner`].
-    #[inline]
-    pub fn inner(&self) -> &StoreInner {
-        &self.pruned.inner
-    }
-
-    /// Returns an exclusive reference to the underlying [`StoreInner`].
-    #[inline]
-    pub fn inner_mut(&mut self) -> &mut StoreInner {
-        &mut self.pruned.inner
-    }
-
-    /// Calls a host `func` at `instance` with `params_results` buffer.
-    ///
-    /// # Errors
-    ///
-    /// If the host function returns an error.
-    pub fn call_host_func(
-        &mut self,
-        func: &HostFuncEntity,
-        instance: Option<&Instance>,
-        params_results: FuncInOut,
-        call_hooks: CallHooks,
-    ) -> Result<(), Error> {
-        self.typed_store()
-            .call_host_func(func, instance, params_results, call_hooks)
-    }
-
-    /// Returns an exclusive reference to [`StoreInner`] and a [`ResourceLimiterRef`].
-    pub fn store_inner_and_resource_limiter_ref(
-        &mut self,
-    ) -> (&mut StoreInner, ResourceLimiterRef) {
-        self.typed_store().store_inner_and_resource_limiter_ref()
-    }
-
-    /// Returns the associated [`TypedStore`] of `self`.
-    fn typed_store(&mut self) -> &mut dyn TypedStore {
-        self.pruned.restore_pruned.clone().restore(self)
-    }
-
-    /// Restores `self` to a proper [`Store<T>`] if possible.
-    ///
-    /// # Errors
-    ///
-    /// If the `T` of the resulting [`Store<T>`] does not match the given `T`.
-    #[inline]
-    fn restore<T: 'static>(&mut self) -> Result<&mut Store<T>, PrunedStoreError> {
-        if unlikely(TypeId::of::<T>() != self.pruned.id) {
-            return Err(PrunedStoreError);
-        }
-        let store = {
-            // Safety: we restore the original `Store<T>` from the pruned `Store<Pruned>`.
-            //
-            // This is safe because we have already checked above that the `TypedId` of `T`
-            // matches the `id` of the original `Store<T>` and thus the `T`'s are identical.
-            //
-            // Furthermore, we are only operating on `&mut` pointers and not values.
-            // Finally, `Store<T>` has the same size and alignment for all `T`.
-            unsafe { mem::transmute::<&mut PrunedStore, &mut Store<T>>(self) }
-        };
-        Ok(store)
-    }
-}
-
-/// Returned when [`PrunedStore::restore`] failed.
-#[derive(Debug)]
-pub struct PrunedStoreError;
-
-#[test]
-fn pruning_works() {
-    let engine = Engine::default();
-    let mut store = Store::new(&engine, ());
-    let pruned = store.prune();
-    assert!(pruned.restore::<()>().is_ok());
-}
-
-#[test]
-fn pruning_errors() {
-    let engine = Engine::default();
-    let mut store = Store::new(&engine, ());
-    let pruned = store.prune();
-    assert!(pruned.restore::<i32>().is_err());
-}
-
-#[test]
-fn pruned_store_deref() {
-    let mut config = crate::Config::default();
-    config.consume_fuel(true);
-    let engine = Engine::new(&config);
-    let mut store = Store::new(&engine, ());
-    let fuel_amount = 100;
-    store.set_fuel(fuel_amount).unwrap();
-    let pruned = store.prune();
-    assert_eq!(
-        PrunedStore::inner(pruned).fuel.get_fuel().unwrap(),
-        fuel_amount
-    );
-    PrunedStore::inner_mut(pruned)
-        .fuel
-        .set_fuel(fuel_amount * 2)
-        .unwrap();
-    assert_eq!(
-        PrunedStore::inner(pruned).fuel.get_fuel().unwrap(),
-        fuel_amount * 2
-    );
 }
 
 /// The inner parts of the [`Store`] which are generic over a host provided `T`.
@@ -1052,7 +892,7 @@ impl<T> Store<T> {
     ///
     /// If fuel metering is disabled.
     pub fn get_fuel(&self) -> Result<u64, Error> {
-        self.inner.fuel.get_fuel().map_err(Into::into)
+        self.inner.fuel.get_fuel().map_err(Error::from)
     }
 
     /// Sets the remaining fuel of the [`Store`] to `value` if fuel metering is enabled.
@@ -1065,7 +905,7 @@ impl<T> Store<T> {
     ///
     /// If fuel metering is disabled.
     pub fn set_fuel(&mut self, fuel: u64) -> Result<(), Error> {
-        self.inner.fuel.set_fuel(fuel).map_err(Into::into)
+        self.inner.fuel.set_fuel(fuel).map_err(Error::from)
     }
 
     /// Allocates a new [`TrampolineEntity`] and returns a [`Trampoline`] reference to it.
